Create compress_tif_folders.sh
This script is designed to be submitted as a slurm job dependent on the successful renaming of fire tifs in the ArrivalDays and FlameLengths folders. The script checks whether the files were renamed (based on the length in characters) and, if they were renamed, compresses all the ArrivalDays, FlameLengths, and ArrivalTimes directories.