Update FSim_post_processing1_merge_fire_tifs_v11_Scenario_refact_of_the_Nordgren.R
There was an error in the merge code where fires were removed from the list of season fires, but then the vectors of information about the fires were the incorrect length. I needed to re-create the vectors of run information from the filtered list of season fires (excluding fires with a burned area of 0).